requirements.txt
bcrypt~=4.2.0
validate-docbr==1.3.0
pymongo~=4.9.1
dotenv
getpass

python-dotenv~=1.0.1
PySimpleGUI~=5.0.6
tkcalendar~=1.6.1
dnspython~=2.6.1